% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSeuratClusterEnrichment.R
\name{testSeuratClusterEnrichment}
\alias{testSeuratClusterEnrichment}
\title{testSeuratClusterEnrichment - Test Seurat Clusters for Enrichment}
\usage{
testSeuratClusterEnrichment(
  markers_list,
  test_set,
  background_set,
  test = "fisher",
  alternative = "greater",
  p_adjust = "bonferroni"
)
}
\arguments{
\item{markers_list}{Data frame. A data frame containing two columns: a
\code{gene} column and a \code{cluster} column. Typically, output from the
FindAllMarkers() function. If using a custom markers_list object, ensure no
clusters are labelled as "0". If so, rename that cluster to prevent erroneous
enrichment. "0" is a cluster label internally reserved for the background set.}

\item{test_set}{Data frame or character. Either a data frame including two columns,
"gene" and "group", or a vector of genes giving the enrichment set.}

\item{background_set}{Character. A vector of genes that form the background
population for the tests.}

\item{test}{Character. One of "fisher" (default) for fisher's exact test, "chi.sq"
for the chi squared test, "hyper" for the hypergeometric test, and "conditional"
for partial odds ratios from a multivariable logistic regression. Conditional
is only applicable when groups are included in the test set and there is substantial
overlap between the groups.}

\item{alternative}{Character. One of "greater" (default), "two.sided", or "less".}

\item{p_adjust}{Character. Correct for multiple testing using one of "holm",
"hochberg", "hommel", "bonferroni" (default), "BH", "BY", or "fdr" methods.}
}
\value{
A data frame with enrichment test results.
}
\description{
Performs appropriate tests of enrichment of Seurat clusters for a specified set
of genes.
}
